fix ofer's PR comments
