+++ app/models/restaurant.rb
@@ -1,15 +1,4 @@
-
-class NameValidator < ActiveModel::Validator
-  def validate(record)
-    if record.blank?
-      record.errors.add(:base, "Name can't be empty")
-      return false
-    elsif Restaurant.where(name: record.name).exists?
-      record.errors.add(:name_exists, "A restaurant with this name already exists")
-      return false
-    end
-  end
-end
+require_relative '../../app/models/validations/restaurant_name_validator'
 
 class Restaurant < ApplicationRecord
   has_many :reviews
